Change `npm run dev` error list format
The previous error output format looked pretty (code-snippets) but didn't allow you to navigate to the error with a simple click.
Switching back to the default format fixes this. The modification to vue.config.js does this.
Updated VSCode tasks and attempted to integrate the `npm run dev` output with the problem matcher. WIP.